code review changes

+++ src/main/java/com/postIT/postIT/NoteResource.java
@@ -8,6 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/note")
@@ -26,12 +27,14 @@ public class NoteResource {
 
     @PostMapping("/add")
     public ResponseEntity<Note> addNote(@RequestBody Note note) {
+        checkLengthOfANote(note);
         Note newNote = noteService.addNote(note);
         return new ResponseEntity<>(newNote, HttpStatus.CREATED);
     }
 
     @PutMapping("/update")
     public ResponseEntity<Note> updateNote(@RequestBody Note note) {
+        checkLengthOfANote(note);
         Note updateNote = noteService.updateNote(note);
         return new ResponseEntity<>(updateNote, HttpStatus.OK);
     }
@@ -39,7 +42,37 @@ public class NoteResource {
     @DeleteMapping("/delete/{id}")
     @Transactional
     public ResponseEntity<?> deleteNote(@PathVariable("id") Long id) {
-        noteService.deleteNoteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (checkIfNotesExists(id)) {
+            noteService.deleteNoteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return ResponseEntity.badRequest().body("Note with this id doesn't exist");
+    }
+
+    private boolean checkIfNotesExists(Long id) {
+        List<Long> notesId = noteService
+                .findAllNotes()
+                .stream()
+                .map(Note::getId)
+                .toList();
+
+        return notesId.contains(id);
+    }
+
+    private void checkLengthOfANote(Note note) {
+        if (note.getContent().length() > 200) {
+            throw new NoteContentExceedsSizeLimitException("Size of note content cannot exceed 200 characters");
+        }
+    }
+
+    @ExceptionHandler(NoteContentExceedsSizeLimitException.class)
+    public ResponseEntity<String> handleNoteContentExceedsSizeLimitException(NoteContentExceedsSizeLimitException e) {
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    public static class NoteContentExceedsSizeLimitException extends RuntimeException {
+        public NoteContentExceedsSizeLimitException(String message) {
+            super(message);
+        }
     }
 }
